feat(ci): roll up "What to Test" across builds at Pre-flight promotion
The weekly Beta→Pre-flight promotion attaches only the newest eligible
build, which carries only its own one-build upload-time "What to Test".
Pre-flight testers promoting weekly never saw the changes accumulated in
the builds that were skipped between promotions.

At promotion, regenerate the promoted build's "What to Test" as a rollup
spanning the deploy/<version> tags of the last Pre-flight build and the
promoted build. Note formatting (conventional-commit parsing, scope and
pattern filtering, char limit) is extracted into changelog.mjs and shared
with set-beta-notes.mjs so upload-time and promotion-time notes read
identically.

Best-effort: a notes failure is swallowed so it can't strand a promoted
build un-submitted. The promote checkout now fetches full history + tags
so git log between the deploy tags resolves.
