+++ .github/workflows/promote-preflight.yml
@@ -24,7 +24,12 @@ jobs:
     name: Evaluate Beta builds and promote latest eligible to Pre-flight
     runs-on: ubuntu-latest
     steps:
+      # Full history + tags: the rollup "What to Test" walks git log between the
+      # deploy/<version> tags of the last Pre-flight build and the promoted build.
       - uses: actions/checkout@v6
+        with:
+          fetch-depth: 0
+          fetch-tags: true
 
       - name: Setup Node
         uses: actions/setup-node@v6
